Attempted to integrate IPC processes
Tried to build the communication process for the main window of the application using a tutorial from https://coursetro.com/posts/code/122/Electron-IPC-Tutorial---Communication-within-your-Electron-App

Not really sure if this is what we need since the tutorial is for transfering data between two windows. We only have one window here, so I think we should be able to bypass all this IPC stuff and send the contents of the array directly to the window.